Add test to assert the endpoints
